Refactor tsx
